+++ src/Atoms/Icon/index.ts
@@ -1,7 +1,9 @@
 import { AddWithCircle } from './components/AddWithCircle';
+import { Archive } from './components/Archive';
 import { ArrowRight } from './components/ArrowRight';
 import { Attachement } from './components/Attachement';
 import { Bank } from './components/Bank';
+import { Bell } from './components/Bell';
 import { Bolt } from './components/Bolt';
 import { CheckMark } from './components/CheckMark';
 import { CheckMarkCircle } from './components/CheckMarkCircle';
@@ -9,8 +11,12 @@ import { ChevronDown } from './components/ChevronDown';
 import { ChevronUp } from './components/ChevronUp';
 import { Cross } from './components/Cross';
 import { CrossCircle } from './components/CrossCircle';
+import { DiagonalWithCircle } from './components/DiagonalWithCircle';
 import { DirectionDown } from './components/DirectionDown';
 import { DirectionUp } from './components/DirectionUp';
+import { ExternalLink } from './components/ExternalLink';
+import { EyeWithPlus } from './components/EyeWithPlus';
+import { EyeWithPlusActive } from './components/EyeWithPlusActive';
 import { File } from './components/File';
 import { Flag } from './components/Flag';
 import { Globe } from './components/Globe';
@@ -32,9 +38,11 @@ import { Rtmt } from './components/Rtmt';
 
 export default {
   AddWithCircle,
+  Archive,
   ArrowRight,
   Attachement,
   Bank,
+  Bell,
   Bolt,
   CheckMark,
   CheckMarkCircle,
@@ -42,8 +50,12 @@ export default {
   ChevronUp,
   Cross,
   CrossCircle,
+  DiagonalWithCircle,
   DirectionDown,
   DirectionUp,
+  ExternalLink,
+  EyeWithPlus,
+  EyeWithPlusActive,
   File,
   Flag,
   Globe,