+++ src/Molecules/FormField/FormField.types.ts
@@ -3,12 +3,15 @@ export type Props = {
   label: string;
   /** But you can hide it visually */
   hideLabel?: boolean;
+
   className?: string;
   error?: string;
   extraInfo?: string;
-  fieldId: string;
-  required?: boolean;
+  forId?: string;
+  group?: boolean;
+  showRequired?: boolean;
   size?: 's';
+
   /**
    * You need to specify width
    * (better in pixels), because