+++ src/Molecules/Input/FormFieldSimple/index.ts
@@ -1,1 +1,0 @@
-export { FormFieldSimple } from './FormFieldSimple';