+++ src/Molecules/Input/index.ts
@@ -1,7 +1,8 @@
 import { Text } from './Text';
 import { Select } from './Select';
 import { Number } from './Number';
+import { Checkbox } from './Checkbox';
 
 // Stitch all inputs so
 // we can use e.g. Input.Text
-export default { Text, Number, Select };
+export default { Text, Number, Select, Checkbox };