+++ src/index.ts
@@ -8,6 +8,7 @@ import Flag from './Atoms/Flag';
 import Flexbox from './Atoms/Flexbox';
 import FormLabel from './Atoms/FormLabel';
 import Icon from './Atoms/Icon';
+import Legend from './Atoms/Legend';
 import List from './Atoms/List';
 import ListItem from './Atoms/ListItem';
 import Media, { useMedia } from './Atoms/Media';
@@ -25,6 +26,8 @@ import Button from './Molecules/Button';
 import CardWithTitle from './Molecules/CardWithTitle';
 import CollapsibleCard from './Molecules/CollapsibleCard';
 import FeedbackBanner from './Molecules/FeedbackBanner';
+import Fieldset from './Molecules/Fieldset';
+import FormField from './Molecules/FormField';
 import Input from './Molecules/Input';
 import LabeledValue from './Molecules/LabeledValue';
 import Link from './Molecules/Link';
@@ -63,12 +66,15 @@ export {
   DateTime,
   Development,
   FeedbackBanner,
+  Fieldset,
   Flag,
   Flexbox,
+  FormField,
   FormLabel,
   Icon,
   Input,
   LabeledValue,
+  Legend,
   Link,
   LinkBuy,
   LinkSell,